timetable/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
requests==2.25.1
pyjwt==2.1.0
